refactor: move message composer hooks to hooks folder's root
