+++ src/components/MessageInput/hooks/index.ts
@@ -1,3 +1,6 @@
+export * from './useAttachmentManagerState';
+export * from './useCanCreatePoll';
 export * from './useCooldownTimer';
 export * from './useMessageInputUiApi';
-export * from './messageComposer';
+export * from './useMessageComposer';
+export * from './useMessageComposerHasSendableData';
