--- src/components/MessageInput/hooks/messageComposer/index.ts
@@ -1,3 +1,0 @@
-export * from './useAttachmentManagerState';
-export * from './useMessageComposerHasSendableData';
-export * from './useMessageComposer';
