+++ src/components/MessageInput/hooks/useSubmitHandler.ts
@@ -1,9 +1,9 @@
 import { useCallback } from 'react';
+import { useMessageComposer } from './useMessageComposer';
 import { useChannelActionContext } from '../../../context/ChannelActionContext';
 import { useTranslationContext } from '../../../context/TranslationContext';
 
 import type { MessageInputProps } from '../MessageInput';
-import { useMessageComposer } from './messageComposer';
 
 export const useSubmitHandler = (props: MessageInputProps) => {
   const { clearEditingState, overrideSubmitHandler } = props;
